backend/requirements.txt
flask
flask-cors
flask_sqlalchemy
python-dotenv
psycopg2-binary